Added a reset button and fixed singleton bug
The reset button should work now without breaking. I had to comment out
the destroyOnLoad because it was giving me a similar error. I included
the link to where I got the advice in GameManager. I just had to tell it
to look for a new instance if none existed.