Added Car Model
I'm still tweaking the human models. There's some slight inconsistencies
with the collision detection and I had to use a bit of a roundabout way
to get the model to stand properly. I imagine there's an easier way, but
I'm not sure. Also put the singleton fix in InputManager to prevent
multiple instances of that being created. Will try to finish up models
and add the UI stuff tomorrow.