Adjusted UI positioning
The buttons on the main scene where overlapping with the win message, so
I tweaked their anchors. I also increased the size of the buttons in the
grid layout group so they would be easier to see and since we never have
enough levels to fill it usually anyways. I also added a "How to Play"
popup to the Title Screen.